Multiple widgets!
Added the ability to provide multiple widgets from one app.
Also made the search for widgets toggleable in build.py (You now need to
call build.py with the '--widget' flag) and hopefully fixed the crashes
that happened if for some reason python or the c libary were not
initialized.