Implemented missing functionality, removed testing package name
Implemented the missing functionality for all Views and Layouts on the
java and the Python side. This makes Chronometers usable and provides
all needed and available functionality. Also fixed a huge derp where all
Layout files depended on my test package name. Additionally, widgets
with a space in their name will now work properly. Added the ability for
LinearLayouts to be horizontal, added a function for widgets to start
the main app and tidied up the code.